requirements.txt
torch
transformers>=4.41.0
sentence-transformers
accelerate
bitsandbytes
peft
nltk
spacy
networkx
numpy
scikit-learn
faiss-cpu
qdrant-client
chromadb
weaviate-client
pymilvus
pandas
Pillow
pydub
opencv-python
redis
openai>=1.0
requests
pydantic
pyyaml
jinja2